data: Move batch review selection translation
Batch selection owns raw display-ID and ownership translation for batch
source lines.

Review-scoped translation also lived there, which made the batch layer
reach upward into persisted file-review state during line actions.

This commit addresses that by moving review-aware batch gutter translation
into data.file_review.batch_selection and routing
include/apply/discard/show-from callers through that boundary.

The batch selection module now stays focused on batch ownership and raw
display helpers.
