+++ src/git_stage_batch/commands/discard_from.py
@@ -15,7 +15,6 @@ from ..batch.selection import (
     resolve_current_batch_binary_file_scope,
     resolve_batch_file_scope,
     require_single_file_context_for_line_selection,
-    translate_batch_file_gutter_ids_to_selection_ids,
     translate_atomic_unit_error_to_gutter_ids,
 )
 from ..batch.submodule_pointer import (
@@ -34,6 +33,7 @@ from ..data.file_review.records import FileReviewAction
 from ..data.file_review.state import (
     resolve_batch_source_action_scope,
 )
+from ..data.file_review.batch_selection import translate_batch_file_gutter_ids_to_selection_ids
 from ..core.buffer import (
     LineBuffer,
     write_buffer_to_path,
